moved grid init to separate file
